+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
 
   * Dependencies updated
+  * Istambul dependencies change to nyc
 
 2.1.0 / 2018-06-10
 ==================