+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
 
 * Using typescript
+* Exports changed to support TypeScript
 
 
 3.0.0 / 2021-08-05